Updated DNABot scripts to OT-2 v2
Note: this requires an additional file thermofisher_96_wellplate_180ul.json to be loaded in through the opentrons app before use. For testing this with opentrons simulator, without having loaded the labware in comment out all of the agar plate transfers. The file and its test are included in this folder.